Fix namespaces for moved files
