+++ DishIndex.Core/Parsers/RecipeParser.cs
@@ -1,7 +1,6 @@
-﻿using DishIndex.Models;
-using DishIndex.Utilities;
+﻿using DishIndex.Core.Models;
+using DishIndex.Core.Utilities;
 using System.Text.RegularExpressions;
-using System.Linq;
 
 namespace DishIndex.Parsers;
