+++ DishIndex.Tests/SerializationTests.cs
@@ -1,7 +1,8 @@
-﻿using DishIndex.Models;
+﻿using DishIndex.Core.Models;
 using S = System.Text.Json;
 using N = Newtonsoft.Json;
 using DishIndex.Core.Generators;
+using DishIndex.Core.Utilities;
 
 namespace DishIndex.Tests;
